# swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    ApiError:
        properties:
            error:
                type: string
                x-go-name: Error
        type: object
        x-go-package: github.com/shahinrahimi/booknest/types
    ApiSuccess:
        properties:
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-package: github.com/shahinrahimi/booknest/types
    Book:
        description: |-
            Book defines the structure for an API book
            swagger: model Book
        properties:
            author:
                description: |-
                    the author for the book
                    exmaple: George Orvil
                type: string
                x-go-name: Author
            cover:
                description: the cover for the book
                example: http://...
                type: string
                x-go-name: Cover
            description:
                description: the description for the book
                example: The story about animals that lives in a farm
                type: string
                x-go-name: Description
            id:
                description: the id for the book
                type: string
                x-go-name: ID
            price:
                description: the price for the book
                example: 10
                format: float
                minimum: 1
                type: number
                x-go-name: Price
            title:
                description: the title for the book
                example: Animal Farm
                type: string
                x-go-name: Title
        required:
            - title
            - author
            - description
            - cover
            - price
        type: object
        x-go-package: github.com/shahinrahimi/booknest/pkg/book
info:
    description: Documentation for [BOOKNEST] API
    title: of [BOOKNEST] API
    version: 1.0.0
paths:
    /api/book/:
        post:
            description: Create a new Book
            operationId: createBook
            responses:
                "201":
                    $ref: '#/responses/successResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - book
    /api/book/{id}:
        delete:
            description: delete the Book
            operationId: deleteBook
            responses:
                "201":
                    $ref: '#/responses/successResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - book
        put:
            description: update the Book
            operationId: updateBook
            responses:
                "201":
                    $ref: '#/responses/successResponse'
                "400":
                    $ref: '#/responses/errorResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - book
    /book:
        get:
            description: Return a list of books from database
            operationId: listBooks
            responses:
                "200":
                    $ref: '#/responses/booksResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - book
    /book/{id}:
        get:
            description: Return a book from database
            operationId: listSingleBook
            responses:
                "200":
                    $ref: '#/responses/bookResponse'
                "404":
                    $ref: '#/responses/errorResponse'
                "500":
                    $ref: '#/responses/errorResponse'
            tags:
                - book
produces:
    - application/json
responses:
    bookResponse:
        description: a single book
        schema:
            $ref: '#/definitions/Book'
    booksResponse:
        description: A list of books
        schema:
            items:
                $ref: '#/definitions/Book'
            type: array
    errorResponse:
        description: ""
        schema:
            $ref: '#/definitions/ApiError'
    successResponse:
        description: ""
        schema:
            $ref: '#/definitions/ApiSuccess'
schemes:
    - http
swagger: "2.0"